// flog/src/components/commons/buttons/outlined/01/OutlinedButton01.types.ts
export interface IOutlinedButton01Props {
  content: string;
  size: string;
  onClick?: () => void;
  disabled?: boolean;
}

export interface IOutlinedButton01UIProps {
  content: string;
  size: string;
  onClick?: () => void;
  disabled?: boolean;
}

export interface IOutlinedButton01UIEmotionProps {
  size: string;
  onClick?: () => void;
  disabled?: boolean;
}